Remove fallback `WebPart`
The promise paradigm guarantees that the reply will only be given when
the event has set the result, thus the default is never used (unless
the thread handling results in an inconsistent state, which should
only happen if there are bugs in the runtime).

This is a breaking change, consumers of the API will have to stop
passing the default result.