Add implementation of remote events
Add an implementation of a "remote event", i.e. two functions that can
be used to connect `Trigger` function to an `IEvent` interface that
internally performs [de]serialization of the event arguments.

This effectively provides a way to create events that can be triggered
and listened to and that can be sent over HTTP.